+++ simplesat/utils.py
